factor out embedly card into its own component; it seems to help
